requirements.txt
requests~=2.31.0
Flask~=2.3.2
WTForms~=3.0.1
python-dotenv~=1.0.0